requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
requests==2.31.0
plotly==5.17.0
seaborn==0.13.0
matplotlib==3.8.2
groq==0.4.2
python-dotenv==1.0.0 